+++ src/test/java/com/twilio/rest/preview/wireless/RatePlanTest.java
@@ -175,4 +175,36 @@ public class RatePlanTest {
 
         RatePlan.updater("WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").update();
     }
+
+    @Test
+    public void testDeleteRequest() {
+        new NonStrictExpectations() {{
+            Request request = new Request(HttpMethod.DELETE,
+                                          Domains.PREVIEW.toString(),
+                                          "/wireless/RatePlans/WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa");
+            
+            twilioRestClient.request(request);
+            times = 1;
+            result = new Response("", 500);
+            twilioRestClient.getAccountSid();
+            result = "AC123";
+        }};
+
+        try {
+            RatePlan.deleter("WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").delete();
+            fail("Expected TwilioException to be thrown for 500");
+        } catch (TwilioException e) {}
+    }
+
+    @Test
+    public void testDeleteResponse() {
+        new NonStrictExpectations() {{
+            twilioRestClient.request((Request) any);
+            result = new Response("null", TwilioRestClient.HTTP_STATUS_CODE_NO_CONTENT);
+            twilioRestClient.getObjectMapper();
+            result = new ObjectMapper();
+        }};
+
+        RatePlan.deleter("WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").delete();
+    }
 }